Removed the old login page and replaced it with new one
